Reland "[graphite] Use SampleCount enum for sample count in TextureInfo"
This reverts commit 19260be441a9c739aa3628275726b8b810c9d0e1.

Reason for revert: staging change landed in Android, no change needed here

Original change's description:
> Revert "[graphite] Use SampleCount enum for sample count in TextureInfo"
>
> This reverts commit cd6d01ab68ff8936b902a91d433bd87b742ca160.
>
> Reason for revert: Blocking Android roll
>
> Original change's description:
> > [graphite] Use SampleCount enum for sample count in TextureInfo
> >
> > Bug: b/452351267
> > Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1100357
> > Commit-Queue: Michael Ludwig <michaelludwig@google.com>
>
> Bug: b/452351267
> No-Presubmit: true
> No-Tree-Checks: true
> No-Try: true
> Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1104856
> Commit-Queue: Robert Phillips <robertphillips@google.com>
> Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>

Bug: b/452351267
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1105816
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Auto-Submit: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>

+++ src/gpu/graphite/dawn/DawnGraphicsPipeline.cpp
@@ -670,7 +670,7 @@ sk_sp<DawnGraphicsPipeline> DawnGraphicsPipeline::Make(
     }
 
     // Multisampled state
-    descriptor.multisample.count = renderPassDesc.fSampleCount;
+    descriptor.multisample.count = (uint8_t) renderPassDesc.fSampleCount;
     descriptor.multisample.mask = 0xFFFFFFFF;
     descriptor.multisample.alphaToCoverageEnabled = false;
